# swagger.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Backend API
  description: The set of API used for backend.lexffe.io. 
servers:
  - url: https://backend.lexffe.io
...
paths:
...
  /coll/:
    get:
      tags: [Collections]
      summary: Gets the complete list of collections
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    type:
                      $ref: "#/components/schemas/ObjectType"
      security:
        - api_key: []
    post:
      tags: [Collections]
      summary: Create a new collection.
      requestBody:
        description: The name and type of the collection.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  description: "Name of the collection"
                  type: string
                type:
                  $ref: "#/components/schemas/ObjectType"
      responses:
        409:
          description: "collection name is in conflict with either router internal routes / existing collections"
        201:
          $ref: "#/components/responses/Created"
        400:
          $ref: "#/components/responses/MalformedReq"
        401:
          $ref: "#/components/responses/UnauthorizedError"
      security:
        - api_key: []
  /coll/{collectionName}/:

    delete:
      tags: [Collections]
      summary: Delete a collection.
      parameters:
        - name: collectionName
          in: path
          description: The name of the collection.
          required: true
          schema:
            type: string
      responses:
        400:
          $ref: "#/components/responses/MalformedReq"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          $ref: "#/components/responses/NotFound"
        204:
          $ref: "#/components/responses/NoContent"
      security:
        - api_key: []
  
  # pages
...
components:
  responses:
    UnauthorizedError:
      description: Unauthorised. (Your token is either invalid, or you did not provide one if the route is private.)
    MalformedReq:
      description: Malformed request. (Usually - JSON request body cannot be binded to model.)
    Created:
      description: Object created.
    NotFound:
      description: The entity you have specified does not exist.
    NoContent:
      description: Successful, with no response.
  schemas:
    ObjectType:
      type: string
      enum: [page, reference]
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer